+++ api/utils/errors/errors.user-not-authorized.js
@@ -1,12 +1,12 @@
 const status = require('http-status')
 
 const defaultMessage = 'User is not allowed'
-const ErrorGeneric = require('./erros.generic_error')
+const ErrorGeneric = require('./erros.generic-error')
 
-module.exports = class ErrorAllowedUser extends ErrorGeneric {
+module.exports = class ErrorNotAuthorized extends ErrorGeneric {
   constructor(message) {
     super(message)
-    Error.captureStackTrace(this, ErrorAllowedUser)
+    Error.captureStackTrace(this, ErrorNotAuthorized)
     this.statusCode = status.FORBIDDEN
     this.message = message || defaultMessage
   }
