+++ api/utils/middlewares/middlewares.validate-dto.js
@@ -1,8 +1,9 @@
 const Joi = require('joi')
-const ErrorBusinessRule = require('../errors/errors.business_rule')
+const ErrorBusinessRule = require('../errors/errors.business-rule')
 
-const validateDTOMiddleware = (type, params, options = {}) => {
-  return async (req, res, next) => {
+const validateDTOMiddleware =
+  (type, params, options = {}) =>
+  async (req, res, next) => {
     const schema = Joi.object().keys(params)
 
     const result = schema.validate(req[type], {
@@ -11,13 +12,13 @@ const validateDTOMiddleware = (type, params, options = {}) => {
     })
 
     if (result.error) {
-      const message = result.error.details.reduce((acc, item) => {
-        return [...acc, item.message]
-      }, [])
+      const message = result.error.details.reduce(
+        (acc, item) => [...acc, item.message],
+        []
+      )
       throw new ErrorBusinessRule(message)
     }
 
     return next()
   }
-}
 module.exports = validateDTOMiddleware
